Move tests around to match file structure
